- Changed SignUpScreen to have everything display in rows (slightly less scuffed).
- Also fixed a bug where when User types in something other than a double in balance, the parsing double failed.

- Added a back button to the sign-up screen.

TODO: Still need to pass and store the meal plan attribute properly. Because right now we ask for residence, but UserGateway stores a list of associated dining halls.

TODO: The sign-up screen is part of the view, and is handling too much logic. Some of its logic should be done in the controller.